Restructured test suite.
