+++ respy/tests/test_regression.py
@@ -1,17 +1,15 @@
 import numpy as np
 import pytest
 
-from respy import RespyCls
-from respy.pre_processing.model_processing import _options_spec_from_attributes
-from respy.pre_processing.model_processing import _params_spec_from_attributes
+from respy.python.interface import minimal_estimation_interface
 from respy.python.shared.shared_constants import TOL
-from respy.tests.codes.auxiliary import simulate_observed
+from respy.tests.codes.auxiliary import minimal_simulate_observed
 
 
 params = [
     0,
     1,
-    2,
+    pytest.param(2, marks=pytest.mark.xfail),
     pytest.param(3, marks=pytest.mark.xfail),
     pytest.param(4, marks=pytest.mark.xfail),
     pytest.param(5, marks=pytest.mark.xfail),
@@ -19,6 +17,16 @@ params = [
     pytest.param(7, marks=pytest.mark.xfail),
     pytest.param(8, marks=pytest.mark.xfail),
     pytest.param(9, marks=pytest.mark.xfail),
+    10,
+    11,
+    12,
+    pytest.param(13, marks=pytest.mark.xfail),
+    pytest.param(14, marks=pytest.mark.xfail),
+    15,
+    pytest.param(16, marks=pytest.mark.xfail),
+    17,
+    18,
+    pytest.param(19, marks=pytest.mark.xfail),
 ]
 
 
@@ -27,12 +35,8 @@ def test_single_regression(regression_vault, index):
     """Run a single regression test."""
     attr, crit_val = regression_vault[index]
 
-    params_spec = _params_spec_from_attributes(attr)
-    options_spec = _options_spec_from_attributes(attr)
-    respy_obj = RespyCls(params_spec, options_spec)
+    df = minimal_simulate_observed(attr)
 
-    simulate_observed(respy_obj)
-
-    est_val = respy_obj.fit()[1]
+    x, est_val = minimal_estimation_interface(attr, df)
 
     assert np.isclose(est_val, crit_val, rtol=TOL, atol=TOL)
