+++ tox.ini
@@ -62,7 +62,7 @@ ignore =
     W503  ; ignores linebreak before binary operator which is enforced by Black
 
 [pytest]
-addopts = --doctest-modules -n auto
+addopts = --doctest-modules -n auto --pdbcls=IPython.terminal.debugger:Pdb
 markers =
     slow: Tests that take a long time to run.
 norecursedirs =
